Citizens will now cross the bridge when built

+++ CitySimulator/Controllers/UiObjects/Status.cpp
@@ -17,37 +17,37 @@ void Status::Init(const int x, const int y)
 {
     this->x = x;
     this->y = y;
-    width = float(WINDOW_WIDTH) / 2 - 255;
+    width = int(float(WINDOW_WIDTH) / 2 - 255);
 
-    rect.setPosition(x, y);
+    rect.setPosition(float(x), float(y));
     rect.setFillColor(WHITE);
     rect.setOutlineColor(BLACK);
-    rect.setOutlineThickness(2);
-    rect.setSize(Vector2f(width, 0));
+    rect.setOutlineThickness(2.f);
+    rect.setSize(Vector2f(float(width), 0.f));
 
     title.setFillColor(BLACK);
     title.setCharacterSize(24);
-    title.setPosition(x + 20, y + 10);
+    title.setPosition(x + 20.f, y + 10.f);
     title.setFont(FontController::Monofur());
 
     buttonText.setFillColor(BLACK);
     buttonText.setCharacterSize(20);
-    buttonText.setPosition(x + 20, y + height - 34);
+    buttonText.setPosition(x + 20.f, y + height - 34.f);
     buttonText.setFont(FontController::Monofur());
     
     content.setFillColor(BLACK);
     content.setCharacterSize(20);
-    content.setPosition(x + 20, y + 42);
+    content.setPosition(x + 20.f, y + 42.f);
     content.setFont(FontController::Monofur());
 
-    button = Button(Vector2f(width - 42, 28), Vector2f(x + 20, y + height - 34), WHITE, MOUSE_OVER_COLOR);
+    button = Button(Vector2f(width - 42.f, 28.f), Vector2f(x + 20.f, y + height - 34.f), WHITE, MOUSE_OVER_COLOR);
     
     closeText.setFillColor(BLACK);
     closeText.setCharacterSize(20);
-    closeText.setPosition(x + width - 28, y + 8);
+    closeText.setPosition(x + width - 28.f, y + 8.f);
     closeText.setFont(FontController::Monofur());
-    CenterString(closeText, "X", x + width - 20);
-    closeButton = Button(Vector2f(20, 20), Vector2f(x + width - 30, y + 10), WHITE, MOUSE_OVER_COLOR);
+    CenterString(closeText, "X", x + width - 20.f);
+    closeButton = Button(Vector2f(20.f, 20.f), Vector2f(x + width - 30.f, y + 10.f), WHITE, MOUSE_OVER_COLOR);
 }
 
 void Status::Draw()
@@ -203,7 +203,7 @@ bool Status::Grow()
     if (changeHeight != height)
     {
         changeHeight += 10;
-        rect.setSize(Vector2f(rect.getSize().x, changeHeight));
+        rect.setSize(Vector2f(float(rect.getSize().x), float(changeHeight)));
         return true;
     }
     return false;
@@ -214,7 +214,7 @@ bool Status::Shrink()
     if (changeHeight != 0)
     {
         changeHeight -= 10;
-        rect.setSize(Vector2f(rect.getSize().x, changeHeight));
+        rect.setSize(Vector2f(float(rect.getSize().x), float(changeHeight)));
         return false;
     }
     return true;
